requirement.txt
Flask~=3.0
Flask-SQLAlchemy~=3.1
SQLAlchemy~=2.0
python-dotenv~=1.0
psycopg2-binary~=2.9
Flask-Bcrypt~=1.0
Flask-JWT-Extended~=4.6

